@@ -40,8 +40,6 @@
     "jest-pnp-resolver": "1.0.1",
     "jest-resolve": "23.6.0",
     "mini-css-extract-plugin": "0.4.3",
-    "mobx": "^4.0.0",
-    "mobx-react": "^5.4.3",
     "moment": "^2.20.1",
     "node-sass": "^4.11.0",
     "optimize-css-assets-webpack-plugin": "5.0.1",
@@ -50,10 +48,10 @@
     "postcss-loader": "3.0.0",
     "postcss-preset-env": "6.3.1",
     "postcss-safe-parser": "4.0.1",
-    "react": "^16.7.0",
+    "react": "^16.8.0",
     "react-app-polyfill": "^0.2.0",
     "react-dev-utils": "^7.0.0",
-    "react-dom": "^16.7.0",
+    "react-dom": "^16.8.0",
     "react-router": "^4.2.0",
     "react-router-dom": "^4.2.2",
     "resolve": "1.8.1",
