@@ -1,37 +1,24 @@
-import React, { Component } from "react";
-import { observer } from "mobx-react";
+import React, { useContext, useState } from "react";
 
-import userStore from "../../stores/userStore";
 import Avatar from "../reusable/Avatar";
 import "./styles/Users.scss";
+import { UserContext } from "../../context/UserContext";
 
-@observer
-class UserSearch extends Component {
-  state = {
+export default function UserSearch({ history }) {
+  const [state, updateState] = useState({
     results: [],
     query: "",
     isLoading: false,
     modalText: null
-  };
+  });
+  const {searchByField} = useContext(UserContext);
 
-  _searchTimeout;
+  const setState = newData => updateState({ ...state, ...newData });
 
-  // searches for users after typing has halted for .3 seconds
-  handleSearchKeyUp = () => {
-    const { query } = this.state;
-    clearTimeout(this._searchTimeout);
-    if (query !== "") {
-      this.setState({ isLoading: true });
-      this._searchTimeout = setTimeout(() => this.searchForUsers(query), 300);
-    } else {
-      this.setState({ results: [], isLoading: false });
-    }
-  };
-
-  searchForUsers = async query => {
+  const searchForUsers = async query => {
     try {
-      const results = await userStore.searchByField(query, "email");
-      this.setState({ results, isLoading: false });
+      const results = await searchByField(query, "email");
+      setState({ results, isLoading: false });
     } catch (error) {
       if (error.message.includes("Failed to fetch")) {
         return prompt(
@@ -42,49 +29,59 @@ class UserSearch extends Component {
     }
   };
 
-  openProfile = user => {
-    this.setState({ query: "", results: [] });
-    this.props.history.push("/profile/" + user.id);
+  // searches for users after typing has halted for .3 seconds
+  const handleSearchKeyUp = () => {
+    let _searchTimeout;
+
+    const { query } = state;
+    clearTimeout(_searchTimeout);
+    if (query !== "") {
+      setState({ isLoading: true });
+      _searchTimeout = setTimeout(() => searchForUsers(query), 300);
+    } else {
+      setState({ results: [], isLoading: false });
+    }
   };
 
-  render() {
-    const { isLoading, query, results } = this.state;
+  const openProfile = user => {
+    setState({ query: "", results: [] });
+    history.push("/profile/" + user.id);
+  };
 
-    return (
-      <div className="UserSearch">
-        <form className="uk-search uk-search-default">
-          <span uk-search-icon="true" uk-icon="icon: search" />
-          <input
-            className="uk-search-input"
-            type="search"
-            value={query}
-            placeholder="Search for users.."
-            onChange={e => this.setState({ query: e.target.value })}
-            onKeyUp={this.handleSearchKeyUp}
-            results={5}
-          />
-          {isLoading && (
-            <div
-              className="uk-form-icon uk-form-icon-flip"
-              uk-spinner="ratio:.5"
-            />
-          )}
-        </form>
-        {results.length > 0 && (
-          <UserSearchResults results={results} openProfile={this.openProfile} />
-        )}
-        {query !== "" && (
-          <span
-            className="onClickOutside"
-            onClick={() => this.setState({ results: [], query: "" })}
+  const { isLoading, query, results } = state;
+
+  return (
+    <div className="UserSearch">
+      <form className="uk-search uk-search-default">
+        <span uk-search-icon="true" uk-icon="icon: search" />
+        <input
+          className="uk-search-input"
+          type="search"
+          value={query}
+          placeholder="Search for users.."
+          onChange={e => setState({ query: e.target.value })}
+          onKeyUp={handleSearchKeyUp}
+          results={5}
+        />
+        {isLoading && (
+          <div
+            className="uk-form-icon uk-form-icon-flip"
+            uk-spinner="ratio:.5"
           />
         )}
-      </div>
-    );
-  }
+      </form>
+      {results.length > 0 && (
+        <UserSearchResults results={results} openProfile={openProfile} />
+      )}
+      {query !== "" && (
+        <span
+          className="onClickOutside"
+          onClick={() => setState({ results: [], query: "" })}
+        />
+      )}
+    </div>
+  );
 }
-
-export default UserSearch;
 
 const UserSearchResults = ({ results, openProfile }) => (
   <div className="uk-card uk-card-default uk-card-body searchResults">
